// src/lib/enums/button.enum.ts
export enum ButtonVariant {
    ELEVATED = 'elevated',
    OUTLINE = 'outline',
    TEXT = 'text',
    GHOST = 'ghost',
}

export enum ButtonShape {
    RECTANGLE = 'rectangle',
    ROUNDED = 'rounded',
    CIRCLE = 'circle',
    SQUARE = 'square',
}

export enum IconPlacement {
    LEADING = 'leading',
    TRAILING = 'trailing',
}